% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_fineres.R
\name{cov_fineres}
\alias{cov_fineres}
\title{Push inputlist to a fine resolution.}
\usage{
cov_fineres(inputFileList, output = "e:/tmp/2.5", targetRes = 2.5)
}
\arguments{
\item{inputFileList}{a character vector specifiying the location of all the input rasters.  Input list should only be a raster type (e.g. .tif).  Best practice is to use list.files(full.names = TRUE).}

\item{output}{destination of the output files.}

\item{targetRes}{desired resolution to convert to.}
}
\description{
Uses raster::disagregate to push a set of rasters to a target resolution.  Note that input raster resolution need to be divisible by the target resolution.
}
\examples{
l <- list.files("e:/covariates/10")
cv_FineRes(l, output = "e:/covariates/2.5")
}
\keyword{disaggregate}
\keyword{raster,}
