% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_tiles.R
\name{cov_tile}
\alias{cov_tile}
\title{Tiles a set of rasters}
\usage{
cov_tile(rList, output = "./tiles/", pxCount = 1000, index = TRUE)
}
\arguments{
\item{rList}{a vector of files to process (e.g. \code{list.files("e:/rasters", pattern= "*.tif")})}

\item{output}{character, destination of output tiles.  Note that a subdirectory for each raster given will be created within this folder.}

\item{pxCount}{the maximum number of pixels in the x and y direction.  \code{pxCount} squared will be the size of the output raster.  Default of 1000 x 1000 pixels.}

\item{index}{When \code{TRUE} genearates a shapefile \code{tiles.shp} -- for each set of tiles generated.}
}
\description{
Uses \code{SpaDES.tools::splitRaster} to generate tiles of a given set of rasters using parallel processing.
}
\examples{
r <- raster::raster(ncol=100, nrow=100) ## empty raster
r <- raster::setValues(r, rnorm(raster::ncell(r), 1)) ##
dir.create("e:/cv_tile")
raster::writeRaster(r, "e:/cv_tile/r.tif")
rasterList <- list.files("e:/cv_tile", full.names = TRUE) ## make sure this is the rasters only!

cv_tile(rList = rasterList, output = "e:/cv_tile",
        pxCount = 50, index = TRUE)
#> [1] "Tiling: r.tif ... 1 of 1"
#> [1] "Generating Tile index: e:/cv_tile/r_tiles/tileindex.gpkg"

list.files("e:/cv_tile", recursive = TRUE, full.names = TRUE)
#> [1] "e:/cv_tile/r.tif"
#> [2] "e:/cv_tile/r_tiles/r_tile1.grd"
#>[10] "e:/cv_tile/r_tiles/tileindex.gpkg"
}
\keyword{index}
\keyword{raster,}
\keyword{tile,}
