% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covariates.R
\name{create_covariates}
\alias{create_covariates}
\title{Covariate generation from input dtm.
\emph{This is an update of cov_dtm.R replacing the processing with 02a} from  BEC_DevExchange_Work_}
\usage{
create_covariates(dtm, SAGApath = "C:/SAGA/", output = "./cv-rasters")
}
\arguments{
\item{dtm}{is a dtm raster object}
}
\description{
Takes a dtm and generates the covariates embeded in this function.
Note that this uses a tools generated by SAGA and as such the correct path to SAGA needs to be specified.
}
\details{
UNDER DEVEOPMENT -- see hack at ~49
}
\examples{
##
create_covariates(dtm,                          ## the dtm (loaded by raster() )
        SAGApath = "C:/SAGA/"         ## specify location of SAGA on your machine
        output   = "c:/dtm-derived" ) ## specify output folder
}
\keyword{SAGA,}
\keyword{covariates,}
\keyword{predictors,}
\keyword{raster}
