% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covariates_plan.R
\name{create_covariates_plan}
\alias{create_covariates_plan}
\title{Intended for creating the sampling plan -- numerous entries have been commented out.
ACTION: perhaps divide these into create_covariates_plan, and create_covariates_model}
\usage{
create_covariates_plan(
  dtm,
  SAGApath = "",
  output = "./cv-rasters",
  layers = "all"
)
}
\arguments{
\item{dtm}{is a dtm raster object}

\item{SAGApath}{Is the location of SAGA on your system.  On linux systems with SAGA GIS installed Use \code{SAGApath = ""}}

\item{output}{Location of where rasters will be saved.}

\item{layers}{The covariates that will be generated.  A full list of covariates is listed at:}
}
\description{
Covariate raster generation from input dtm.
}
\details{
\emph{This is an update of cov_dtm.R replacing the processing with 02a} from  BEC_DevExchange_Work_

Takes a dtm and generates the covariates embeded in this function via SAGA GIS.
This script has been tested with SAGA 7.3 on Windows and Ubuntu 18.
Depending on your system the path to \code{saga_cmd} may need to be specified.
}
\examples{
##
create_covariates(dtm,                          ## the dtm (loaded by raster() )
        SAGApath = "C:/SAGA/"         ## specify location of SAGA on your machine
        output   = "c:/dtm-derived" ) ## specify output folder
}
\keyword{SAGA,}
\keyword{covariates,}
\keyword{predictors,}
\keyword{raster}
