% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_dtm.R
\name{cv_dtm}
\alias{cv_dtm}
\title{Covariate generation from input dtm}
\usage{
cv_dtm(dtm, SAGApath = "C:/SAGA/", output = "./cv-rasters")
}
\arguments{
\item{dtm}{is a dtm raster object}
}
\description{
Takes a dtm and generates the covariates embeded in this function.
Note that this uses a tools generated by SAGA and as such the correct path to SAGA needs to be specified.
}
\details{
UNDER DEVEOPMENT -- see hack at ~49
}
\examples{
## Load sf object
}
\keyword{SAGA,}
\keyword{covariates,}
\keyword{predictors,}
\keyword{raster}
