% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gen.R
\name{model_gen}
\alias{model_gen}
\title{Generate a machine learning model}
\usage{
model_gen(traindat, target, outDir = ".", rseed = NA)
}
\arguments{
\item{traindat}{Is a dataframe that contains the model training data.  The reponse variable should be one of the columns.}

\item{target}{The name of the response variable in the traindat data frame.}

\item{outDir}{Highly recommended to be set as an absolute directory.  This defaults to the project's root directory OR where the RMD script is saved.
Additional products generated from the associated<tt>model_gen.Rmd</tt> markdown script will also be saved to this dir.}

\item{rseed}{Optional random number seed.}
}
\description{
This function takes in all the data needed to produce machine learning model.
Inputs are handed to a RMD report/ script.
Outputs include the markdown report, the cross validation object,
and a binary model (RDS) that can then be used to predict on new data.
}
\examples{
dat <- read.csv("e:/workspace/2020/PEM/ALRF_PEMv2/dev/modDat.csv",
               stringsAsFactors = TRUE)


model_gen(traindat = dat,
          target = "SiteSeries",
          outDir = "e:/tmp/model_gen_test",
          rseed = 456)
}
\keyword{machine-learning,}
\keyword{model,}
\keyword{report}
