% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_landscape.R
\name{predict_landscape}
\alias{predict_landscape}
\title{Predict Landscape}
\usage{
predict_landscape(model, cov, tilesize = 500, outDir = "./predicted")
}
\arguments{
\item{model}{Location of a \code{mlr} model object (i.e. path to it.)}

\item{cov}{A list of raster files.  These will be loaded as a \code{stars} object}

\item{tilesize}{Specify the number of pixels in the \code{x} and \code{y} directions for the tiles to be generated.  If your computer is mememory limited use a smaller tile (e.g. 500).}

\item{outDir}{directory for the output file.}
}
\description{
Takes a mlr model object and associated covariate rasters to generate a thematic map.
In order to process the landscape level prediction the input co-variated are tiled
and then mosaic'd together at the end.
}
\details{
\strong{Action} _is there a bug in tilemaker or with stars causing a xy offset error.
Solution is currently a hack.  See \verb{load tile area}.
}
\examples{
### Testing
cov <- list.files("e:/tmpGIS/PEM_cvs/", pattern = "*.tif",full.names = TRUE)
cov <- cov[-(grep(cov, pattern = "xml"))]

predict_landscape(model = "e:/tmp/model_gen_test/model.rds",
                  cov = cov,
                  tilesize = 1000,
                  outDir = "e:/tmp/predict_landscape")
}
\keyword{landscape}
\keyword{predict,}
