% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_avenza.R
\name{r_avenza}
\alias{r_avenza}
\title{r_avenza}
\usage{
r_avenza(shp, gpx)
}
\arguments{
\item{shp}{The shapefile data loaded as an \emph{sf} object.}

\item{gpx}{The gpx data loaded as an \emph{sf} object.}

\item{crs}{NOT IMPLEMENTED / PENDING The projection string or epsg code for the output data to transformed to.}

\item{att}{NOT IMPLEMENTED / PENDING ... Additional attributes or data columns you wish to keep.  For example, custom attributes fields generated in Avenza.  By default, the point id, name, time, Photos, and desc fields are kept.}
}
\description{
Loads data colleted in the ipad app \emph{Avenza Maps} arranges it by
time stamp and returns an organized sf data frame.
Requires that Avenza data is exported in GPX and SHP formats.
This is done as the GPX data provides an detailed timestamp.
Where in the SHP data provides easier to manage attribute data.
}
\examples{
gpsData <- r_avenza(shpData, gpxData,
                    crs = 3157,
                    attributes = c("")
}
\keyword{Avenza,}
\keyword{GPS,}
\keyword{import}
