% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_erase.R
\name{st_erase}
\alias{st_erase}
\title{Erase Overlapping area of a polygon}
\usage{
st_erase(x, y)
}
\arguments{
\item{x}{an sf layer.  This is the layer that you would like to remove from.}

\item{y}{An sf polygon layer.  This area that will be removed from polygon x.  Obviously only areas that are overlapping will be removed}
}
\description{
Analogous to QGIS' clip tool.  This function is taken from sf documentation.
Layers should be of the same class and projection
}
\details{
z = x - y  -- where z is the resulting polygon
}
\keyword{area}
\keyword{clip,}
\keyword{erase,}
\keyword{remove}
