% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects_sample.R
\name{transect_sample}
\alias{transect_sample}
\title{Transect Sample}
\usage{
transect_sample(lines, mdist = 5)
}
\arguments{
\item{lines}{is a sf \emph{LINES} object.}

\item{mdist}{Optional with a default of 5m}
}
\description{
This is a wrapper for st_line_sample.  The PEM transect lines are used to generate a minimum of one point per line segment. Additional sample points are added at a minimum distance of 5 metres.
}
\details{
\emph{Note: current version does not include the raster method.}
}
\keyword{export}
\keyword{points,}
\keyword{samples,}
\keyword{sampling}
\keyword{transects,}
