% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gen_tidy.R
\name{model_gen_tidy}
\alias{model_gen_tidy}
\title{Generate a machine learning model using tidy models}
\usage{
model_gen_tidy(
  trDat,
  target,
  target2,
  tid,
  outDir = ".",
  mname = "model",
  rseed = NA,
  infiles = infiles,
  mmu = mmu,
  field_transect = field_transect
)
}
\arguments{
\item{trDat}{Is a dataframe that contains the model training data.  The reponse variable should be one of the columns.}

\item{target}{The name of the response variable in the traindat data frame.}

\item{outDir}{Highly recommended to be set as an absolute directory.  This defaults to the project's root directory OR where the RMD script is saved.
Additional products generated from the associated `model_gen_tidy.Rmd`` markdown script will also be saved to this dir.}

\item{rseed}{Optional random number seed.}
}
\description{
This function takes in all the data needed to produce machine learning model.
Inputs are handed to a RMD report/ script.
Outputs include the markdown report, the cross validation object,
and a binary model (RDS) that can then be used to predict on new data.
}
\examples{


}
\keyword{machine-learning,}
\keyword{model,}
\keyword{report}
